Changed name of project to 'DroughtCast'
