+++ dc/R/PlotError.R
