+++ dc/R/ReadSMAPProducts.R
