+++ dc/fetch/FetchGridmet.py
@@ -1,4 +1,4 @@
-from crdm.fetch.DownloadFile import download_file
+from dc.fetch.DownloadFile import download_file
 import os
 from pathlib import Path
