+++ dc/fetch/FetchL4SM_L4C.py
@@ -1,4 +1,4 @@
-from crdm.fetch.DownloadFile import download_file
+from dc.fetch.DownloadFile import download_file
 import os
 import datetime as dt
 from pathlib import Path
